moved functors into traits class to replace using directives
