+++ Test/expression_optimizer.cpp
@@ -141,7 +141,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_matrix_vector_expression_optimize ){
 	//diagonal-matrix-product
 	{
 		typedef diagonal_matrix<V1> M;
-		typedef functors::scalar_binary_multiply<double> F;
+		typedef device_traits<cpu_tag>::multiply<double> F;
 		V1 v1 = create_vector(5);
 		V v = create_vector(5);
 		M m(v1);
@@ -249,7 +249,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_matrix_row_optimize ){
 	//matrix unary
 	{
 		M1 m1 = create_matrix(5,10);
-		typedef functors::scalar_sqr<double> F;
+		typedef device_traits<cpu_tag>::sqr<double> F;
 		vector_unary<matrix_row<M1 const>, F> e1 = row(sqr(m1),1);
 		vector_unary<matrix_row<matrix_transpose<M1 const> >,F> e2 = column(sqr(m1),1);
 		BOOST_CHECK_SMALL(norm_inf(e1 - row(matrix<double>(sqr(m1)),1)), 1.e-10);
@@ -259,7 +259,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_matrix_row_optimize ){
 	{
 		M1 m1 = create_matrix(5,10);
 		M2 m2 = create_matrix(5,10);
-		typedef functors::scalar_binary_multiply<double> F;
+		typedef device_traits<cpu_tag>::multiply<double> F;
 		vector_binary<matrix_row<M1 const>,matrix_row<M2 const>,F> e1 = row(m1*m2,1);
 		vector_binary<matrix_row<matrix_transpose<M1 const>>,matrix_row<matrix_transpose<M2 const> >,F> e2 = column(m1*m2,1);
 		BOOST_CHECK_SMALL(norm_inf(e1 - row(matrix<double>(m1*m2),1)), 1.e-10);
@@ -309,7 +309,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_vector_range_optimize ){
 	//vector unary
 	{
 		V1 v1 = create_vector(10);
-		typedef functors::scalar_sqr<double> F;
+		typedef device_traits<cpu_tag>::sqr<double> F;
 		vector_unary<vector_range<V1 const>, F> e1 = subrange(sqr(v1),1,4);
 		BOOST_CHECK_SMALL(norm_inf(e1 - subrange(vector<double>(sqr(v1)),1,4)), 1.e-10);
 	}
@@ -317,7 +317,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_vector_range_optimize ){
 	{
 		V1 v1 = create_vector(10);
 		V2 v2 = create_vector(10);
-		typedef functors::scalar_binary_multiply<double> F;
+		typedef device_traits<cpu_tag>::multiply<double> F;
 		vector_binary<vector_range<V1 const>,vector_range<V2 const>,F> e1 = subrange(v1*v2,1,4);
 		BOOST_CHECK_SMALL(norm_inf(e1 - subrange(vector<double>(v1*v2),1,4)), 1.e-10);
 	}
@@ -361,7 +361,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_matrix_range_optimize ){
 	//matrix unary
 	{
 		M1 m1 = create_matrix(5,10);
-		typedef functors::scalar_sqr<double> F;
+		typedef device_traits<cpu_tag>::sqr<double> F;
 		matrix_unary<matrix_range<M1 const>, F> e1 = subrange(sqr(m1),1,4,3,7);
 		BOOST_CHECK_SMALL(norm_inf(e1 - subrange(matrix<double>(sqr(m1)),1,4,3,7)), 1.e-10);
 	}
@@ -369,7 +369,7 @@ BOOST_AUTO_TEST_CASE( Remora_prod_matrix_range_optimize ){
 	{
 		M1 m1 = create_matrix(5,10);
 		M2 m2 = create_matrix(5,10);
-		typedef functors::scalar_binary_multiply<double> F;
+		typedef device_traits<cpu_tag>::multiply<double> F;
 		matrix_binary<matrix_range<M1 const>,matrix_range<M2 const>,F> e1 = subrange(m1*m2,1,4,3,7);
 		BOOST_CHECK_SMALL(norm_inf(e1 - subrange(matrix<double>(m1*m2),1,4,3,7)), 1.e-10);
 	}
